// @generated/user/user.model.d.ts
import { Role } from '../prisma/role.enum';
import { StorageMember } from '../storage-member/storage-member.model';
import { Log } from '../log/log.model';
import { Project } from '../project/project.model';
import { ConstructionLog } from '../construction-log/construction-log.model';
import { LogComment } from '../log-comment/log-comment.model';
import { LogCommentReply } from '../log-comment-reply/log-comment-reply.model';
import { PermitInspection } from '../permit-inspection/permit-inspection.model';
import { MaterialEntry } from '../material-entry/material-entry.model';
import { DiscartItem } from '../discart-item/discart-item.model';
import { UserCount } from './user-count.output';
export declare class User {
    id: number;
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    nickname: string | null;
    public: boolean;
    role: keyof typeof Role;
    emailVerified: boolean;
    profilePicture: string | null;
    about: string | null;
    createdAt: Date;
    updatedAt: Date;
    isPremium: boolean;
    subscriptionId: string | null;
    expiresAt: Date | null;
    willExpireAt: Date | null;
    apartment: string | null;
    isApprovedResident: boolean;
    isAdmin: boolean;
    storageMemberships?: Array<StorageMember>;
    logs?: Array<Log>;
    projectResponsible?: Array<Project>;
    constructionLogs?: Array<ConstructionLog>;
    logComments?: Array<LogComment>;
    logCommentReplies?: Array<LogCommentReply>;
    responsibleInspections?: Array<PermitInspection>;
    receivedMaterials?: Array<MaterialEntry>;
    discartItems?: Array<DiscartItem>;
    _count?: UserCount;
}